+++ LastResort/ParticleGMissile.cpp
@@ -19,8 +19,8 @@ void Particle_G_Missile::Move() {
 void Particle_G_Missile::Draw()
 {
 	if (going_up == true || go_left == true)
-		App->render->Blit(texture, position.x, position.y - anim.GetCurrentFrame().h / 2, &anim.GetFrame());
+		App->render->Blit(texture, (int)position.x, (int)position.y - anim.GetCurrentFrame().h / 2, &anim.GetFrame());
 	else
-		App->render->BlitEx(texture, position.x, position.y - anim.GetCurrentFrame().h / 2, &anim.GetFrame(), SDL_FLIP_HORIZONTAL);
+		App->render->BlitEx(texture, (int)position.x, (int)position.y - anim.GetCurrentFrame().h / 2, &anim.GetFrame(), SDL_FLIP_HORIZONTAL);
 
 }
